Add winCondition
"Win Conditions" refer to things that directly lead to the protagonists
dying and the mastermind winning. They do not attempt to outline the
path by which the mastermind could arrive at such an outcome. Not only
is it fairly difficult to write code to do that, it isn't exactly in the
spirit of good fun. The game is a challenge between the mastermind and
the protagonists - not a challenge between the protagonists and an
all-knowing, ever-winning machine.